requirements.txt
python-docx>=1.2.0
docxtpl>=0.20.0
lxml>=6.0.0
olefile>=0.47
pymupdf>=1.27.0
pytesseract>=0.3.10
Pillow>=10.0.0
